:zap: 후기 삭제(관리자용) JPA API 구현

+++ src/main/java/com/farmer/cornfarmer/src/admin/AdminController.java
@@ -155,7 +155,7 @@ public class AdminController {
      */
     @ResponseBody
     @DeleteMapping("/reviews/{reviewIdx}")
-    public BaseResponse deleteReview(@PathVariable int reviewIdx) {
+    public BaseResponse deleteReview(@PathVariable long reviewIdx) {
         try {
             //admin 계정인지 검증
             adminService.deleteReview(reviewIdx);
